+++ VolviewPlugIns/vvITKThresholdSegmentationLevelSet.cxx
@@ -28,7 +28,7 @@ class ThresholdSegmentationLevelSeteRunner
       module.SetUpdateMessage("Computing Segmentation Threshold Level Set...");
       module.ProcessData( pds  );
       sprintf(tmp,"Total number of iterations = %d \n Final RMS error = %g",
-                         module.GetFilter()->GetElapsedIterations(),
+                         static_cast< int >( module.GetFilter()->GetElapsedIterations() ),
                          module.GetFilter()->GetRMSChange());
       info->SetProperty( info, VVP_REPORT_TEXT, tmp );
